+++ yazi-adapter/src/drivers/kgp_old.rs
@@ -1,15 +1,17 @@
 use core::str;
 use std::{io::Write, path::PathBuf};
 
-use anyhow::Result;
+use anyhow::{Result, bail};
 use base64::{Engine, engine::general_purpose};
 use image::DynamicImage;
 use ratatui_core::layout::Rect;
-use yazi_emulator::{CLOSE, ESCAPE, Emulator, START};
+use yazi_emulator::{CLOSE, EMULATOR, ESCAPE, Emulator, START};
+use yazi_ffi::shm::NamedSharedMemory;
 use yazi_macro::writef;
 use yazi_tty::{TTY, sequence::MoveTo};
 
-use crate::{ADAPTOR, Image, drivers::Kgp};
+use super::KgpPayload;
+use crate::{ADAPTOR, Image, drivers::kgp_id};
 
 pub(super) struct KgpOld;
 
@@ -34,44 +36,63 @@ impl KgpOld {
 	pub(super) fn image_erase(area: Rect) -> Result<()> {
 		let mut w = TTY.lockout();
 		let Some(shown) = ADAPTOR.shown_area() else {
-			writef!(w, "{START}_Gq=2,a=d,d=I,i={}{ESCAPE}\\{CLOSE}", Kgp::image_id())?;
+			writef!(w, "{START}_Gq=2,a=d,d=I,i={}{ESCAPE}\\{CLOSE}", kgp_id())?;
 			return Ok(());
 		};
 
 		for y in area.top()..area.bottom() {
 			for x in area.left()..area.right() {
 				let p = (y - shown.y) as u32 * shown.width as u32 + (x - shown.x) as u32 + 1;
-				write!(w, "{START}_Gq=2,a=d,d=i,i={},p={p}{ESCAPE}\\{CLOSE}", Kgp::image_id())?;
+				write!(w, "{START}_Gq=2,a=d,d=i,i={},p={p}{ESCAPE}\\{CLOSE}", kgp_id())?;
 			}
 		}
 		w.flush()?;
 		Ok(())
 	}
 
-	async fn encode(img: DynamicImage, size: (u32, u32)) -> Result<Vec<u8>> {
-		fn output(raw: &[u8], format: u8, (w, h): (u32, u32)) -> Result<Vec<u8>> {
-			let b64 = general_purpose::STANDARD.encode(raw).into_bytes();
+	async fn encode(img: DynamicImage, size: (u32, u32)) -> Result<KgpPayload> {
+		fn output(raw: &[u8], format: u8, size: (u32, u32)) -> Result<KgpPayload> {
+			output_shm(raw, format, size).or_else(|_| output_b64(raw, format, size))
+		}
 
+		fn output_shm(raw: &[u8], format: u8, (w, h): (u32, u32)) -> Result<KgpPayload> {
+			if !EMULATOR.kgp_shm.get() {
+				bail!("Shared memory is not supported by the terminal")
+			}
+
+			let mut pl = KgpPayload::with(200, NamedSharedMemory::new(raw)?);
+			write!(
+				pl,
+				"{START}_Gq=2,a=t,t=s,i={},f={format},s={w},v={h},S={};{}{ESCAPE}\\{CLOSE}",
+				kgp_id(),
+				raw.len(),
+				pl.name(),
+			)?;
+
+			Ok(pl)
+		}
+
+		fn output_b64(raw: &[u8], format: u8, (w, h): (u32, u32)) -> Result<KgpPayload> {
+			let b64 = general_purpose::STANDARD.encode(raw).into_bytes();
 			let mut it = b64.chunks(4096).peekable();
-			let mut buf = Vec::with_capacity(b64.len() + it.len() * 50);
+			let mut pl = KgpPayload::new(b64.len() + it.len() * 50);
 			if let Some(first) = it.next() {
 				write!(
-					buf,
+					pl,
 					"{START}_Gq=2,a=t,i={},f={format},s={w},v={h},m={};{}{ESCAPE}\\{CLOSE}",
-					Kgp::image_id(),
+					kgp_id(),
 					it.peek().is_some() as u8,
 					unsafe { str::from_utf8_unchecked(first) },
 				)?;
 			}
 
 			while let Some(chunk) = it.next() {
-				write!(buf, "{START}_Gm={};{}{ESCAPE}\\{CLOSE}", it.peek().is_some() as u8, unsafe {
+				write!(pl, "{START}_Gm={};{}{ESCAPE}\\{CLOSE}", it.peek().is_some() as u8, unsafe {
 					str::from_utf8_unchecked(chunk)
 				})?;
 			}
 
-			write!(buf, "{CLOSE}")?;
-			Ok(buf)
+			Ok(pl)
 		}
 
 		tokio::task::spawn_blocking(move || match img {
@@ -98,7 +119,7 @@ impl KgpOld {
 					buf,
 					"{}{START}_Gq=2,a=p,i={},p={p},x={left},y={top},w={},h={},c=1,r=1,z=-1,C=1{ESCAPE}\\{CLOSE}",
 					MoveTo(area.x + x as u16, area.y + y as u16),
-					Kgp::image_id(),
+					kgp_id(),
 					right - left,
 					bottom - top,
 				)?;